Modularize reports
Entry pages are kept within the main view, but each report now gets
its own module which might consist typically of:

 * A tabular report (default)
 * A graphical version of the same report (optional)

Plus its own helpers where they are not shared across other modules.

Standard templates (non-report web pages) have been prettified a
little to help with code review and any future refactors.
