+++ AIPscan/Reporter/templates/view_copy.html
@@ -5,8 +5,26 @@
 <h3>Preservation copy: {{ copy.name }}</h3>
 
 <table class="table table-bordered table-condensed" style="margin-top: 20px;">
-  <tr><td width=20%><strong>Original file</strong></td><td><a href="{{ url_for('reporter.view_original', id=original.id ) }}"">{{ original.name }}</td></tr>
-  <tr><td width=20%><strong>AIP</strong></td><td><a href="{{ url_for('reporter.view_aip', id=aip.id ) }}"">{{ aip.transfer_name }}  {{ aip.uuid }}</a></td></tr>
+  <tr>
+  	<td width=20%>
+  		<strong>Original file</strong>
+  	</td>
+  	<td>
+  		<a href="{{ url_for('reporter.view_original', original_file_id=original.id ) }}"">
+  			{{ original.name }}
+  		</a>
+  	</td>
+  </tr>
+  <tr>
+  	<td width=20%>
+  		<strong>AIP</strong>
+  	</td>
+  	<td>
+  		<a href="{{ url_for('reporter.view_aip', aip_id=aip.id ) }}"">
+  			{{ aip.transfer_name }}  {{ aip.uuid }}
+  		</a>
+  	</td>
+  </tr>
   <tr><td width=20%><strong>UUID</strong></td><td>{{ copy.uuid }}</td></tr>
   <tr><td width=20%><strong>Normalization date</strong></td><td>{{ copy.normalization_date }}</td></tr>
   <tr><td width=20%><strong>Size</strong></td><td>{{ size }}</td></tr>
